+++ examples/with-next-js/pages/_app.js
@@ -1,5 +1,5 @@
-import '../app.css'
+import '../app.css';
 
 export default function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return <Component {...pageProps} />;
 }